Extract `to_ref_or_refs` implementation into submodule
