+++ cedar-policy-core/src/parser/cst_to_ast.rs
@@ -34,14 +34,14 @@
 // cloning.
 
 use super::cst;
-use super::err::{self, parse_errors, ParseError, ParseErrors, ToASTError, ToASTErrorKind};
+use super::err::{parse_errors, ParseError, ParseErrors, ToASTError, ToASTErrorKind};
 use super::loc::Loc;
 use super::node::Node;
 use super::unescape::{to_pattern, to_unescaped_string};
 use super::util::{flatten_tuple_2, flatten_tuple_3, flatten_tuple_4};
 use crate::ast::{
-    self, ActionConstraint, CallStyle, EntityReference, EntityUID, Integer, PatternElem,
-    PolicySetError, PrincipalConstraint, PrincipalOrResourceConstraint, ResourceConstraint,
+    self, ActionConstraint, CallStyle, Integer, PatternElem, PolicySetError, PrincipalConstraint,
+    PrincipalOrResourceConstraint, ResourceConstraint,
 };
 use crate::est::extract_single_argument;
 use itertools::Either;
@@ -50,6 +50,12 @@ use std::cmp::Ordering;
 use std::collections::{BTreeMap, HashSet};
 use std::mem;
 use std::sync::Arc;
+
+/// Defines the function `cst::Expr::to_ref_or_refs` and other similar functions
+/// for converting CST expressions into one or multiple entity UIDS. Used to
+/// extract entity uids from expression which appear in the policy scope.
+mod to_ref_or_refs;
+use to_ref_or_refs::OneOrMultipleRefs;
 
 /// Type alias for convenience
 type Result<T> = std::result::Result<T, ParseErrors>;
@@ -810,33 +816,6 @@ impl ExprOrSpecial<'_> {
 }
 
 impl Node<Option<cst::Expr>> {
-    fn to_ref(&self, var: ast::Var) -> Result<EntityUID> {
-        self.to_ref_or_refs::<SingleEntity>(var).map(|x| x.0)
-    }
-
-    fn to_ref_or_slot(&self, var: ast::Var) -> Result<EntityReference> {
-        self.to_ref_or_refs::<EntityReference>(var)
-    }
-
-    fn to_refs(&self, var: ast::Var) -> Result<OneOrMultipleRefs> {
-        self.to_ref_or_refs::<OneOrMultipleRefs>(var)
-    }
-
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let expr = self.try_as_inner()?;
-
-        match &*expr.expr {
-            cst::ExprData::Or(o) => o.to_ref_or_refs::<T>(var),
-            cst::ExprData::If(_, _, _) => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "an `if` expression",
-                    None::<String>,
-                ))
-                .into()),
-        }
-    }
-
     /// convert `cst::Expr` to `ast::Expr`
     pub fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
@@ -861,111 +840,6 @@ impl Node<Option<cst::Expr>> {
     }
 }
 
-/// Type level marker for parsing sets of entity uids or single uids
-/// This presents having either a large level of code duplication
-/// or runtime data.
-/// This marker is (currently) only used for translating entity references
-/// in the policy scope.
-trait RefKind: Sized {
-    fn err_str() -> &'static str;
-    fn create_single_ref(e: EntityUID, loc: &Loc) -> Result<Self>;
-    fn create_multiple_refs(es: Vec<EntityUID>, loc: &Loc) -> Result<Self>;
-    fn create_slot(loc: &Loc) -> Result<Self>;
-}
-
-struct SingleEntity(pub EntityUID);
-
-impl RefKind for SingleEntity {
-    fn err_str() -> &'static str {
-        "an entity uid"
-    }
-
-    fn create_single_ref(e: EntityUID, _loc: &Loc) -> Result<Self> {
-        Ok(SingleEntity(e))
-    }
-
-    fn create_multiple_refs(_es: Vec<EntityUID>, loc: &Loc) -> Result<Self> {
-        Err(ToASTError::new(
-            ToASTErrorKind::wrong_entity_argument_one_expected(
-                err::parse_errors::Ref::Single,
-                err::parse_errors::Ref::Set,
-            ),
-            loc.clone(),
-        )
-        .into())
-    }
-
-    fn create_slot(loc: &Loc) -> Result<Self> {
-        Err(ToASTError::new(
-            ToASTErrorKind::wrong_entity_argument_one_expected(
-                err::parse_errors::Ref::Single,
-                err::parse_errors::Ref::Template,
-            ),
-            loc.clone(),
-        )
-        .into())
-    }
-}
-
-impl RefKind for EntityReference {
-    fn err_str() -> &'static str {
-        "an entity uid or matching template slot"
-    }
-
-    fn create_slot(_loc: &Loc) -> Result<Self> {
-        Ok(EntityReference::Slot)
-    }
-
-    fn create_single_ref(e: EntityUID, _loc: &Loc) -> Result<Self> {
-        Ok(EntityReference::euid(Arc::new(e)))
-    }
-
-    fn create_multiple_refs(_es: Vec<EntityUID>, loc: &Loc) -> Result<Self> {
-        Err(ToASTError::new(
-            ToASTErrorKind::wrong_entity_argument_two_expected(
-                err::parse_errors::Ref::Single,
-                err::parse_errors::Ref::Template,
-                err::parse_errors::Ref::Set,
-            ),
-            loc.clone(),
-        )
-        .into())
-    }
-}
-
-/// Simple utility enum for parsing lists/individual entityuids
-#[derive(Debug)]
-enum OneOrMultipleRefs {
-    Single(EntityUID),
-    Multiple(Vec<EntityUID>),
-}
-
-impl RefKind for OneOrMultipleRefs {
-    fn err_str() -> &'static str {
-        "an entity uid or set of entity uids"
-    }
-
-    fn create_slot(loc: &Loc) -> Result<Self> {
-        Err(ToASTError::new(
-            ToASTErrorKind::wrong_entity_argument_two_expected(
-                err::parse_errors::Ref::Single,
-                err::parse_errors::Ref::Set,
-                err::parse_errors::Ref::Template,
-            ),
-            loc.clone(),
-        )
-        .into())
-    }
-
-    fn create_single_ref(e: EntityUID, _loc: &Loc) -> Result<Self> {
-        Ok(OneOrMultipleRefs::Single(e))
-    }
-
-    fn create_multiple_refs(es: Vec<EntityUID>, _loc: &Loc) -> Result<Self> {
-        Ok(OneOrMultipleRefs::Multiple(es))
-    }
-}
-
 impl Node<Option<cst::Or>> {
     fn to_expr_or_special(&self) -> Result<ExprOrSpecial<'_>> {
         let or = self.try_as_inner()?;
@@ -984,39 +858,9 @@ impl Node<Option<cst::Or>> {
             }),
         }
     }
-
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let or = self.try_as_inner()?;
-
-        match or.extended.len() {
-            0 => or.initial.to_ref_or_refs::<T>(var),
-            _n => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "a `||` expression",
-                    None::<String>,
-                ))
-                .into()),
-        }
-    }
 }
 
 impl Node<Option<cst::And>> {
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let and = self.try_as_inner()?;
-
-        match and.extended.len() {
-            0 => and.initial.to_ref_or_refs::<T>(var),
-            _n => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "a `&&` expression",
-                    None::<String>,
-                ))
-                .into()),
-        }
-    }
-
     fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
     }
@@ -1040,44 +884,6 @@ impl Node<Option<cst::And>> {
 }
 
 impl Node<Option<cst::Relation>> {
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let rel = self.try_as_inner()?;
-
-        match rel {
-            cst::Relation::Common { initial, extended } => match extended.len() {
-                0 => initial.to_ref_or_refs::<T>(var),
-                _n => Err(self
-                    .to_ast_err(ToASTErrorKind::wrong_node(
-                        T::err_str(),
-                        "a binary operator",
-                        None::<String>,
-                    ))
-                    .into()),
-            },
-            cst::Relation::Has { .. } => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "a `has` expression",
-                    None::<String>,
-                ))
-                .into()),
-            cst::Relation::Like { .. } => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "a `like` expression",
-                    None::<String>,
-                ))
-                .into()),
-            cst::Relation::IsIn { .. } => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "an `is` expression",
-                    None::<String>,
-                ))
-                .into()),
-        }
-    }
-
     fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
     }
@@ -1158,17 +964,6 @@ impl Node<Option<cst::Relation>> {
 }
 
 impl Node<Option<cst::Add>> {
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let add = self.try_as_inner()?;
-
-        match add.extended.len() {
-            0 => add.initial.to_ref_or_refs::<T>(var),
-            _n => {
-                Err(self.to_ast_err(ToASTErrorKind::wrong_node(T::err_str(), "a `+/-` expression", Some("entity types and namespaces cannot use `+` or `-` characters -- perhaps try `_` or `::` instead?"))).into())
-            }
-        }
-    }
-
     fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
     }
@@ -1196,21 +991,6 @@ impl Node<Option<cst::Add>> {
 }
 
 impl Node<Option<cst::Mult>> {
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let mult = self.try_as_inner()?;
-
-        match mult.extended.len() {
-            0 => mult.initial.to_ref_or_refs::<T>(var),
-            _n => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "a `*` expression",
-                    None::<String>,
-                ))
-                .into()),
-        }
-    }
-
     fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
     }
@@ -1243,21 +1023,6 @@ impl Node<Option<cst::Mult>> {
 }
 
 impl Node<Option<cst::Unary>> {
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let unary = self.try_as_inner()?;
-
-        match &unary.op {
-            Some(op) => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    format!("a `{op}` expression"),
-                    None::<String>,
-                ))
-                .into()),
-            None => unary.item.to_ref_or_refs::<T>(var),
-        }
-    }
-
     fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
     }
@@ -1348,17 +1113,6 @@ impl Node<Option<cst::Member>> {
         match m.item.as_ref().node.as_ref()? {
             cst::Primary::Literal(lit) => lit.as_ref().node.as_ref(),
             _ => None,
-        }
-    }
-
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let mem = self.try_as_inner()?;
-
-        match mem.access.len() {
-            0 => mem.item.to_ref_or_refs::<T>(var),
-            _n => {
-                Err(self.to_ast_err(ToASTErrorKind::wrong_node(T::err_str(), "a `.` expression", Some("entity types and namespaces cannot use `.` characters -- perhaps try `_` or `::` instead?"))).into())
-            }
         }
     }
 
@@ -1587,72 +1341,6 @@ impl Node<Option<cst::MemAccess>> {
 }
 
 impl Node<Option<cst::Primary>> {
-    fn to_ref_or_refs<T: RefKind>(&self, var: ast::Var) -> Result<T> {
-        let prim = self.try_as_inner()?;
-
-        match prim {
-            cst::Primary::Slot(s) => {
-                // Call `create_slot` first so that we fail immediately if the
-                // `RefKind` does not permit slots, and only then complain if
-                // it's the wrong slot. This avoids getting an error
-                // `found ?action instead of ?action` when `action` doesn't
-                // support slots.
-                let slot_ref = T::create_slot(&self.loc)?;
-                let slot = s.try_as_inner()?;
-                if slot.matches(var) {
-                    Ok(slot_ref)
-                } else {
-                    Err(self
-                        .to_ast_err(ToASTErrorKind::wrong_node(
-                            T::err_str(),
-                            format!("{slot} instead of ?{var}"),
-                            None::<String>,
-                        ))
-                        .into())
-                }
-            }
-            cst::Primary::Literal(lit) => {
-                let found = match lit.as_inner() {
-                    Some(lit) => format!("literal `{lit}`"),
-                    None => "empty node".to_string(),
-                };
-                Err(self
-                    .to_ast_err(ToASTErrorKind::wrong_node(
-                        T::err_str(),
-                        found,
-                        None::<String>,
-                    ))
-                    .into())
-            }
-            cst::Primary::Ref(x) => T::create_single_ref(x.to_ref()?, &self.loc),
-            cst::Primary::Name(name) => {
-                let found = match name.as_inner() {
-                    Some(name) => format!("name `{name}`"),
-                    None => "name".to_string(),
-                };
-                Err(self
-                    .to_ast_err(ToASTErrorKind::wrong_node(
-                        T::err_str(),
-                        found,
-                        None::<String>,
-                    ))
-                    .into())
-            }
-            cst::Primary::Expr(x) => x.to_ref_or_refs::<T>(var),
-            cst::Primary::EList(lst) => {
-                let v = ParseErrors::transpose(lst.iter().map(|expr| expr.to_ref(var)))?;
-                T::create_multiple_refs(v, &self.loc)
-            }
-            cst::Primary::RInits(_) => Err(self
-                .to_ast_err(ToASTErrorKind::wrong_node(
-                    T::err_str(),
-                    "record initializer",
-                    None::<String>,
-                ))
-                .into()),
-        }
-    }
-
     pub(crate) fn to_expr(&self) -> Result<ast::Expr> {
         self.to_expr_or_special()?.into_expr()
     }
@@ -2129,7 +1817,7 @@ fn construct_expr_record(kvs: Vec<(SmolStr, ast::Expr)>, loc: Loc) -> Result<ast
 mod tests {
     use super::*;
     use crate::{
-        ast::Expr,
+        ast::{EntityUID, Expr},
         parser::{err::ParseErrors, test_utils::*, *},
         test_utils::*,
     };
